containers/python/requirements.txt
selenium
bs4
pandas
psycopg2-binary